Extract remote publishing processing into service. Leave only task related code in task implementation.
